+++ draft-js-plugins-editor/src/Editor/__test__/index.js
@@ -165,6 +165,10 @@ describe('Editor', () => {
       const expectedSecondArgument = {
         getEditorState: pluginEditor.getEditorState,
         setEditorState: pluginEditor.onChange,
+        getPlugins: pluginEditor.getPlugins,
+        getProps: pluginEditor.getProps,
+        getReadOnly: pluginEditor.getReadOnly,
+        setReadOnly: pluginEditor.setReadOnly,
       };
       draftEditor.props.handleKeyCommand('command');
       expect(plugin.handleKeyCommand).has.been.calledOnce();
@@ -291,6 +295,10 @@ describe('Editor', () => {
       const expectedSecondArgument = {
         getEditorState: pluginEditor.getEditorState,
         setEditorState: pluginEditor.onChange,
+        getPlugins: pluginEditor.getPlugins,
+        getProps: pluginEditor.getProps,
+        getReadOnly: pluginEditor.getReadOnly,
+        setReadOnly: pluginEditor.setReadOnly,
       };
       draftEditor.props.blockRendererFn('command');
       expect(plugin.blockRendererFn).has.been.calledOnce();
@@ -460,36 +468,6 @@ describe('Editor', () => {
       draftEditor.props.blockRendererFn();
       expect(plugin.blockRendererFn).has.been.called();
       expect(customHook).has.been.called();
-    });
-
-    it('renders block component using blockRenderFn prop and decorators', () => {
-      const decorator = (Comp) => (props) => <div className="decorator"><Comp {...props} /></div>;
-      const component = () => null;
-
-      const plugin = {
-        blockRendererFn: () => ({
-          decorators: [decorator],
-          props: { pluginProp: true },
-        }),
-      };
-
-      customHook = () => ({
-        component,
-        props: { editorProp: true },
-      });
-
-      const wrapper = mount(
-        <PluginEditor
-          editorState={ editorState }
-          onChange={ onChange }
-          plugins={ [plugin] }
-          blockRendererFn={ customHook }
-        />
-      );
-
-      const decorators = wrapper.findWhere(n => n.hasClass('decorator'));
-      expect(decorators.length).to.equal(1);
-      expect(wrapper.find(component).length).to.equal(1);
     });
   });