+++ draft-js-plugins-editor/CHANGELOG.md
@@ -5,6 +5,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## To Be Released
 
+### Added
+
+- `onChange` & and all handlers now also receive: `getPlugins`, `getProps`, `getReadOnly`, `setReadOnly`.
+
+### Removed
+
+- `decorators` don't decorate plugins anymore.
+
 ## 1.1.0 - 2016-05-30
 
 ### Added